editor: extract headless TextEditorCore (refactor C1)
The 5400-LOC TextEditor monolith splits into two layers:

- `TextEditorCore.{h,cc}` (476 + 2030 LOC, new) — the headless
  editing substrate. Owns the text buffer, cursor/selection state,
  undo/redo, colorize pipeline, and every editing operation. Zero
  ImGui dependencies. Builds as a standalone library
  `//donner/editor:text_editor_core` with only `//donner/base` and
  `:text_buffer` in deps.
- `TextEditor.{h,cc}` (1173 + 2693 LOC, shrank from 1307 + 4130) —
  the ImGui shell. Owns the rendering pipeline (all `renderXxx`),
  keyboard/mouse input capture, layout math (coordinate↔screen
  conversion, scroll handling), autocomplete popup, snippet engine,
  find/replace UI hooks, and the shortcut table. Depends on
  `:text_editor_core` + `@imgui`.

The shell keeps reference-aliases in its constructor's member-init
list binding short-name fields (`text_`, `state_`, `palette_`, etc.)
to the core's storage. This preserves every render/input call site
without a mechanical rename across 500+ references — a classic
strangler-fig approach to large refactors.

Four shell hooks let the core invoke shell-side ImGui helpers from
inside editing methods:
- `ensureCursorVisibleHook` → TextEditor::ensureCursorVisible
- `requestAutocompleteHook` → flips requestAutocomplete_ + readyFor
- `functionTooltipHook` → TextEditor::handleFunctionTooltip
- `onContentUpdateInternal` → fires the public onContentUpdate cb

`Palette` redefined as `std::array<unsigned int, N>` (ImU32 is
`typedef unsigned int`, so the types are identical in practice).
This keeps the core free of `#include "imgui.h"` while still
interoperating with every ImGui drawlist call in the shell.

Preserves the baseline exactly: 60 tests pass / 13 tests fail in
`text_editor_tests`, matching pre-refactor. The 13 failures expose
real bugs that will be easier to fix now that they live in a
narrower file. Every other editor test target passes unchanged.

Still in the shell (not yet moved):
- `copy`/`cut`/`paste` — waiting on the C2 clipboard abstraction
  integration commit
- `processFind`/`processReplace` — inline ImGui::SetKeyboardFocusHere
  blocks a clean move; defer until another ImGui seam is introduced
- `getCorrectCursorPosition` — only called by the shell's tooltip code
- `handleFunctionTooltip` — legitimate shell-only concern

UB fix: `updateFoldColumn` previously used
`foldEnd_[&fold - &foldBegin_[0]]` which is UB when `folds` is
`foldEnd_` (pointer arithmetic crosses unrelated arrays). Replaced
with an explicit matching-end parameter derived by index. No test
in the 73-test suite exercises this path; the fix is opportunistic.

Friend-class test declarations still live on TextEditor — they get
removed in C3 when the tests migrate to the headless core.
