+++ donner/editor/TextEditor.cc
@@ -37,7 +37,28 @@ constexpr bool IsAscii(char c) {
 
 }  // namespace
 
-TextEditor::TextEditor() {
+TextEditor::TextEditor()
+    : text_(core_.buffer()),
+      state_(core_.mutableState()),
+      replaceIndex_(core_.replaceIndexRef()),
+      foldBegin_(core_.foldBegin()),
+      foldEnd_(core_.foldEnd()),
+      foldSorted_(core_.foldSorted()),
+      scrollbarMarkers_(core_.scrollbarMarkersRef()),
+      changedLines_(core_.changedLines()),
+      tabSize_(core_.tabSizeRef()),
+      scrollToCursor_(core_.scrollToCursorRef()),
+      scrollToTop_(core_.scrollToTopRef()),
+      cursorPositionChanged_(core_.cursorPositionChangedRef()),
+      textChanged_(core_.textChangedRef()),
+      autoIndentOnPaste_(core_.autoIndentOnPasteRef()),
+      selectionMode_(core_.selectionModeRef()),
+      paletteBase_(core_.mutablePaletteBase()),
+      palette_(core_.mutablePalette()),
+      languageDefinition_(core_.getLanguageDefinition()),
+      errorMarkers_(core_.mutableErrorMarkers()),
+      interactiveStart_(core_.interactiveStart()),
+      interactiveEnd_(core_.interactiveEnd()) {
   // Initialize with default dark color scheme
   setPalette(getDarkPalette());
   setLanguageDefinition(LanguageDefinition::SVG());
@@ -49,6 +70,25 @@ TextEditor::TextEditor() {
   startTime_ = std::chrono::duration_cast<std::chrono::milliseconds>(
                    std::chrono::system_clock::now().time_since_epoch())
                    .count();
+
+  // Wire the core's shell-hook callbacks so that moved editing paths
+  // (e.g. `backspace`, `enterCharacter`) can still invoke the ImGui-side
+  // helpers that live in this shell. When the shell is driven from a
+  // headless test context (C3), these hooks remain set but the inner
+  // methods early-return outside `withinRender_`.
+  core_.ensureCursorVisibleHook = [this] { ensureCursorVisible(); };
+  core_.requestAutocompleteHook = [this] {
+    requestAutocomplete_ = true;
+    readyForAutocomplete_ = false;
+  };
+  core_.functionTooltipHook = [this](char32_t ch, const Coordinates& pos) {
+    handleFunctionTooltip(static_cast<ImWchar>(ch), pos);
+  };
+  core_.onContentUpdateInternal = [this] {
+    if (onContentUpdate) {
+      onContentUpdate(this);
+    }
+  };
 }
 
 TextEditor::~TextEditor() = default;
@@ -148,123 +188,45 @@ std::vector<Shortcut> TextEditor::getDefaultShortcuts() {
   return shortcuts;
 }
 
+// ---------------------------------------------------------------------------
+// Forwarders to TextEditorCore. The bodies live in TextEditorCore.cc; the
+// shell keeps the public API for source-level compatibility.
+// ---------------------------------------------------------------------------
+
 void TextEditor::setLanguageDefinition(const LanguageDefinition& langDef) {
-  languageDefinition_ = langDef;
-  regexList_.clear();
-
-  // Build optimized regex list for syntax highlighting
-  for (const auto& regex : languageDefinition_.tokenRegexStrings) {
-    regexList_.emplace_back(std::regex(regex.first, std::regex_constants::optimize), regex.second);
-  }
-
-  colorize();
+  core_.setLanguageDefinition(langDef);
 }
 
 void TextEditor::setPalette(const Palette& value) {
-  paletteBase_ = value;
+  core_.setPalette(value);
 }
 
 std::string TextEditor::getText(const Coordinates& start, const Coordinates& end) const {
-  return text_.getText(start, end);
+  return core_.getText(start, end);
 }
 
 Coordinates TextEditor::getActualCursorCoordinates() const {
-  return sanitizeCoordinates(state_.cursorPosition);
+  return core_.getActualCursorCoordinates();
 }
 
 Coordinates TextEditor::sanitizeCoordinates(const Coordinates& value) const {
-  int line = value.line;
-  int column = value.column;
-
-  // Handle out of bounds line numbers
-  if (line >= text_.getTotalLines()) {
-    line = text_.getTotalLines() - 1;
-    column = text_.getLineMaxColumn(line);
-    return Coordinates(line, column);
-  }
-
-  // Clamp column to valid range
-  column = std::max(0, std::min(column, text_.getLineMaxColumn(line)));
-
-  return Coordinates(line, column);
+  return core_.sanitizeCoordinates(value);
 }
 
 void TextEditor::advance(Coordinates& coords) const {
-  if (coords.line >= text_.getTotalLines()) {
-    return;
-  }
-
-  const Line& line = text_.getLineGlyphs(coords.line);
-  const auto charIndex = text_.getCharacterIndex(coords);
-
-  if (charIndex + 1 < static_cast<int>(line.size())) {
-    // Get length of current UTF-8 sequence at position
-    int delta = Utf8::SequenceLength(line[charIndex].character);
-    const int newIndex = std::min(charIndex + delta, static_cast<int>(line.size()) - 1);
-    coords.column = text_.getCharacterColumn(coords.line, newIndex);
-  } else {
-    // Move to start of next line
-    ++coords.line;
-    coords.column = 0;
-  }
+  core_.advance(coords);
 }
 
 void TextEditor::deleteRange(const Coordinates& start, const Coordinates& end) {
-  text_.deleteRange(start, end);
-
-  // Update scrollbar markers if enabled
-  if (scrollbarMarkers_) {
-    for (int i = 0; i < changedLines_.size(); i++) {
-      if (changedLines_[i] > end.line) {
-        changedLines_[i] -= (end.line - start.line);
-      } else if (changedLines_[i] > start.line && changedLines_[i] < end.line) {
-        changedLines_.erase(changedLines_.begin() + i);
-        i--;
-      }
-    }
-  }
-
-  textChanged_ = true;
-  if (onContentUpdate) {
-    onContentUpdate(this);
-  }
+  core_.deleteRange(start, end);
 }
 
-int TextEditor::insertTextAt(Coordinates& /* inout */ where, std::string_view text, bool indent) {
-  // Insert the text using TextBuffer, which handles basic insertion
-  const int totalLines = text_.insertTextAt(where, text, indent);
-
-  // Handle fold markers which TextBuffer doesn't manage
-  for (size_t i = 0; i < text.length(); ++i) {
-    if (text[i] == '{') {
-      foldBegin_.push_back(where);
-      foldSorted_ = false;
-    } else if (text[i] == '}') {
-      foldEnd_.push_back(where);
-      foldSorted_ = false;
-    }
-  }
-
-  // Track changes for scrollbar markers
-  if (scrollbarMarkers_) {
-    if (std::find(changedLines_.begin(), changedLines_.end(), where.line) == changedLines_.end()) {
-      changedLines_.push_back(where.line);
-    }
-  }
-
-  // Set flags and trigger callback
-  textChanged_ = true;
-  if (onContentUpdate) {
-    onContentUpdate(this);
-  }
-
-  return totalLines;
+int TextEditor::insertTextAt(Coordinates& where, std::string_view text, bool indent) {
+  return core_.insertTextAt(where, text, indent);
 }
 
 void TextEditor::addUndo(UndoRecord& record) {
-  undoBuffer_.resize(static_cast<size_t>(undoIndex_ + 1));
-  undoBuffer_.back() = record;
-  ++undoIndex_;
+  core_.addUndo(record);
 }
 
 namespace {
@@ -402,326 +364,47 @@ Coordinates TextEditor::mousePosToCoordinates(const ImVec2& position) const {
 }
 
 Coordinates TextEditor::findWordStart(const Coordinates& from) const {
-  if (from.line >= text_.getTotalLines()) {
-    return from;
-  }
-
-  const Line& line = text_.getLineGlyphs(from.line);
-  const int charIndex = text_.getCharacterIndex(from);
-
-  if (charIndex >= static_cast<int>(line.size())) {
-    return from;
-  }
-
-  // Skip trailing spaces
-  int currentIndex = charIndex;
-  while (currentIndex > 0 && std::isspace(line[currentIndex].character)) {
-    --currentIndex;
-  }
-
-  // Find word boundary using syntax highlighting and space boundaries
-  const auto startColor = static_cast<ColorIndex>(line[currentIndex].colorIndex);
-  while (currentIndex > 0) {
-    const char c = line[currentIndex].character;
-    if ((c & 0xC0) != 0x80) {  // Not UTF-8 continuation byte
-      if (c <= 32 && std::isspace(c)) {
-        currentIndex++;
-        break;
-      }
-      if (startColor != static_cast<ColorIndex>(line[currentIndex - 1].colorIndex)) {
-        break;
-      }
-    }
-    --currentIndex;
-  }
-
-  return Coordinates(from.line, text_.getCharacterColumn(from.line, currentIndex));
+  return core_.findWordStart(from);
 }
 
 Coordinates TextEditor::findWordEnd(const Coordinates& from) const {
-  if (from.line >= text_.getTotalLines()) {
-    return from;
-  }
-
-  const Line& line = text_.getLineGlyphs(from.line);
-  const int charIndex = text_.getCharacterIndex(from);
-
-  if (charIndex >= static_cast<int>(line.size())) {
-    return from;
-  }
-
-  int currentIndex = charIndex;
-  const bool prevSpace = std::isspace(line[currentIndex].character);
-  const auto startColor = static_cast<ColorIndex>(line[currentIndex].colorIndex);
-
-  while (currentIndex < static_cast<int>(line.size())) {
-    const char c = line[currentIndex].character;
-    const int charLen = Utf8::SequenceLength(c);
-
-    if (startColor != static_cast<ColorIndex>(line[currentIndex].colorIndex)) {
-      break;
-    }
-
-    const bool currSpace = std::isspace(c);
-    if (prevSpace != currSpace) {
-      if (currSpace) {
-        while (currentIndex < static_cast<int>(line.size()) &&
-               std::isspace(line[currentIndex].character)) {
-          ++currentIndex;
-        }
-      }
-      break;
-    }
-    currentIndex += charLen;
-  }
-
-  return Coordinates(from.line, text_.getCharacterColumn(from.line, currentIndex));
+  return core_.findWordEnd(from);
 }
 
 Coordinates TextEditor::findNextWord(const Coordinates& from) const {
-  if (from.line >= text_.getTotalLines()) {
-    return from;
-  }
-
-  Coordinates current = from;
-  int charIndex = text_.getCharacterIndex(from);
-  bool isWord = false;
-  bool skip = false;
-
-  // Check initial character
-  if (current.line < text_.getTotalLines()) {
-    const Line& line = text_.getLineGlyphs(current.line);
-    if (charIndex < static_cast<int>(line.size())) {
-      isWord = std::isalnum(line[charIndex].character);
-      skip = isWord;
-    }
-  }
-
-  // Find next word start
-  while (!isWord || skip) {
-    if (current.line >= text_.getTotalLines()) {
-      const int lastLine = std::max(0, text_.getTotalLines() - 1);
-      return Coordinates(lastLine, text_.getLineMaxColumn(lastLine));
-    }
-
-    const Line& line = text_.getLineGlyphs(current.line);
-    if (charIndex < static_cast<int>(line.size())) {
-      isWord = std::isalnum(line[charIndex].character);
-
-      if (isWord && !skip) {
-        return Coordinates(current.line, text_.getCharacterColumn(current.line, charIndex));
-      }
-
-      if (!isWord) {
-        skip = false;
-      }
-
-      charIndex++;
-    } else {
-      charIndex = 0;
-      ++current.line;
-      skip = false;
-      isWord = false;
-    }
-  }
-
-  return current;
+  return core_.findNextWord(from);
 }
 
 bool TextEditor::isOnWordBoundary(const Coordinates& at) const {
-  if (at.line >= text_.getTotalLines() || at.column == 0) {
-    return true;
-  }
-
-  const Line& line = text_.getLineGlyphs(at.line);
-  const int charIndex = text_.getCharacterIndex(at);
-
-  if (charIndex >= static_cast<int>(line.size())) {
-    return true;
-  }
-
-  if (colorizerEnabled_) {
-    return line[charIndex].colorIndex != line[charIndex - 1].colorIndex;
-  }
-
-  return std::isspace(line[charIndex].character) != std::isspace(line[charIndex - 1].character);
+  return core_.isOnWordBoundary(at);
 }
 
 void TextEditor::removeLine(int start, int end) {
-  UTILS_RELEASE_ASSERT(end >= start);
-  UTILS_RELEASE_ASSERT(text_.getTotalLines() > static_cast<size_t>(end - start));
-
-  // Update error markers
-  ErrorMarkers updatedMarkers;
-  for (const auto& marker : errorMarkers_) {
-    int line = marker.first >= start ? marker.first - 1 : marker.first;
-    if (line >= start && line <= end) {
-      continue;
-    }
-    updatedMarkers.insert({line, marker.second});
-  }
-  errorMarkers_ = std::move(updatedMarkers);
-
-  // Remove lines from buffer
-  text_.removeLine(start, end + 1);
-
-  // Update scrollbar markers
-  if (scrollbarMarkers_) {
-    for (int i = 0; i < static_cast<int>(changedLines_.size()); i++) {
-      if (changedLines_[i] > end) {
-        changedLines_[i] -= (end - start);
-      } else if (changedLines_[i] >= start && changedLines_[i] <= end) {
-        changedLines_.erase(changedLines_.begin() + i);
-        i--;
-      }
-    }
-  }
-
-  textChanged_ = true;
-  if (onContentUpdate) {
-    onContentUpdate(this);
-  }
+  core_.removeLine(start, end);
 }
 
 void TextEditor::removeLine(int index) {
-  UTILS_RELEASE_ASSERT(text_.getTotalLines() > 1);
-
-  // Update error markers
-  ErrorMarkers updatedMarkers;
-  for (const auto& marker : errorMarkers_) {
-    int line = marker.first > index ? marker.first - 1 : marker.first;
-    if (line - 1 == index) {
-      continue;
-    }
-    updatedMarkers.insert({line, marker.second});
-  }
-  errorMarkers_ = std::move(updatedMarkers);
-
-  // Remove line from buffer
-  text_.removeLine(index);
-
-  // Remove folds
-  removeFolds(Coordinates(index, 0), Coordinates(index, 100000));
-
-  // Update scrollbar markers
-  if (scrollbarMarkers_) {
-    for (int i = 0; i < static_cast<int>(changedLines_.size()); i++) {
-      if (changedLines_[i] > index) {
-        changedLines_[i]--;
-      } else if (changedLines_[i] == index) {
-        changedLines_.erase(changedLines_.begin() + i);
-        i--;
-      }
-    }
-  }
-
-  textChanged_ = true;
-  if (onContentUpdate) {
-    onContentUpdate(this);
-  }
+  core_.removeLine(index);
 }
 
 Line& TextEditor::insertLine(int index, int column) {
-  Line& result = text_.insertLine(index, column);
-
-  // Update fold positions
-  for (auto& fold : foldBegin_) {
-    if (fold.line > index - 1 || (fold.line == index - 1 && fold.column >= column)) {
-      fold.line++;
-    }
-  }
-
-  for (auto& fold : foldEnd_) {
-    if (fold.line > index - 1 || (fold.line == index - 1 && fold.column >= column)) {
-      fold.line++;
-    }
-  }
-
-  // Update error markers
-  ErrorMarkers updatedMarkers;
-  for (const auto& marker : errorMarkers_) {
-    updatedMarkers.insert({marker.first >= index ? marker.first + 1 : marker.first, marker.second});
-  }
-  errorMarkers_ = std::move(updatedMarkers);
-
-  return result;
+  return core_.insertLine(index, column);
 }
 
 RcString TextEditor::getWordUnderCursor() const {
-  const Coordinates coords(getCursorPosition().line, std::max(getCursorPosition().column - 1, 0));
-  return getWordAt(coords);
+  return core_.getWordUnderCursor();
 }
 
 RcString TextEditor::getWordAt(const Coordinates& coords) const {
-  const Coordinates start = findWordStart(coords);
-  const Coordinates end = findWordEnd(coords);
-
-  std::string result;
-  const Line& line = text_.getLineGlyphs(coords.line);
-  const int startIndex = text_.getCharacterIndex(start);
-  const int endIndex = text_.getCharacterIndex(end);
-
-  for (int i = startIndex; i < endIndex; ++i) {
-    result.push_back(line[i].character);
-  }
-
-  return RcString(result);
+  return core_.getWordAt(coords);
 }
 
 ImU32 TextEditor::getGlyphColor(const Glyph& glyph) const {
-  if (!colorizerEnabled_) {
-    return palette_[static_cast<int>(ColorIndex::Default)];
-  }
-
-  if (glyph.isComment) {
-    return palette_[static_cast<int>(ColorIndex::Comment)];
-  }
-
-  if (glyph.isMultiLineComment) {
-    return palette_[static_cast<int>(ColorIndex::MultiLineComment)];
-  }
-
-  return palette_[static_cast<int>(glyph.colorIndex)];
+  return core_.getGlyphColor(glyph);
 }
 
 Coordinates TextEditor::findFirst(std::string_view searchText, const Coordinates& start) const {
-  const int totalLines = text_.getTotalLines();
-
-  if (start.line < 0 || start.line >= totalLines) {
-    return Coordinates(totalLines, 0);
-  }
-
-  const std::string text = getText(start, Coordinates(totalLines, 0));
-  const std::string_view textView = text;
-
-  size_t index = 0;
-  size_t found = StringUtils::Find(textView, searchText);
-  Coordinates current = start;
-
-  while (found != std::string::npos) {
-    // Advance to found position
-    while (index < found) {
-      if (textView[index] == '\n') {
-        current.column = 0;
-        current.line++;
-      } else {
-        current.column++;
-      }
-      index++;
-    }
-
-    // Convert character index to column position
-    current.column = text_.getCharacterColumn(current.line, current.column);
-
-    // Check if this is a word boundary match
-    if (getWordAt(current) == searchText) {
-      return current;
-    }
-
-    found = textView.find(searchText, found + 1);
-  }
-
-  return Coordinates(totalLines, 0);
+  return core_.findFirst(searchText, start);
 }
 
 void TextEditor::handleCharacterInput(const ImGuiIO& io, bool shift, bool& keepAutocompleteOpen,
@@ -1395,45 +1078,7 @@ void TextEditor::renderSelection(int lineNo, const Line& line, const ImVec2& lin
 
 void TextEditor::setSelection(const Coordinates& start, const Coordinates& end,
                               SelectionMode mode) {
-  const auto oldSelStart = state_.selectionStart;
-  const auto oldSelEnd = state_.selectionEnd;
-
-  state_.selectionStart = sanitizeCoordinates(start);
-  state_.selectionEnd = sanitizeCoordinates(end);
-
-  if (state_.selectionStart > state_.selectionEnd) {
-    std::swap(state_.selectionStart, state_.selectionEnd);
-  }
-
-  switch (mode) {
-    case SelectionMode::Normal: break;
-
-    case SelectionMode::Word: {
-      state_.selectionStart = findWordStart(state_.selectionStart);
-      if (!isOnWordBoundary(state_.selectionEnd)) {
-        state_.selectionEnd = findWordEnd(findWordStart(state_.selectionEnd));
-      }
-      break;
-    }
-
-    case SelectionMode::Line: {
-      const auto lineNo = state_.selectionEnd.line;
-      state_.selectionStart = Coordinates(state_.selectionStart.line, 0);
-      state_.selectionEnd = Coordinates(lineNo, text_.getLineMaxColumn(lineNo));
-      break;
-    }
-  }
-
-  if (state_.selectionStart != oldSelStart || state_.selectionEnd != oldSelEnd) {
-    cursorPositionChanged_ = true;
-  }
-
-  // Update replace index for find/replace functionality
-  replaceIndex_ = 0;
-  for (size_t ln = 0; ln < state_.cursorPosition.line; ln++) {
-    replaceIndex_ += text_.getLineCharacterCount(ln) + 1;
-  }
-  replaceIndex_ += state_.cursorPosition.column;
+  core_.setSelection(start, end, mode);
 }
 
 namespace {
@@ -1821,90 +1466,12 @@ void TextEditor::openFunctionDeclarationTooltip(std::string_view declaration,
 }
 
 void TextEditor::removeFolds(const Coordinates& start, const Coordinates& end) {
-  removeFolds(foldBegin_, start, end);
-  removeFolds(foldEnd_, start, end);
+  core_.removeFolds(start, end);
 }
 
 void TextEditor::removeFolds(std::vector<Coordinates>& folds, const Coordinates& start,
                              const Coordinates& end) {
-  const bool deleteFullyLastLine = end.line >= text_.getTotalLines() || end.column >= 100000;
-
-  for (int i = 0; i < static_cast<int>(folds.size()); i++) {
-    if (folds[i].line < start.line || folds[i].line > end.line) {
-      if (folds[i].line > end.line) {
-        folds[i].line -= (end.line - start.line) + deleteFullyLastLine;
-      }
-      continue;
-    }
-
-    // Handle fold on start line
-    if (folds[i].line == start.line && start.line != end.line) {
-      if (folds[i].column >= start.column) {
-        folds.erase(folds.begin() + i);
-        foldSorted_ = false;
-        i--;
-      }
-      continue;
-    }
-
-    // Handle fold on end line
-    if (folds[i].line == end.line) {
-      if (folds[i].column < end.column) {
-        if (end.line != start.line || folds[i].column >= start.column) {
-          folds.erase(folds.begin() + i);
-          foldSorted_ = false;
-          i--;
-        }
-      } else {
-        if (end.line == start.line) {
-          folds[i].column = std::max(0, folds[i].column - (end.column - start.column));
-        } else {
-          updateFoldColumn(folds[i], start, end);
-          folds[i].line -= (end.line - start.line);
-        }
-      }
-      continue;
-    }
-
-    // Handle folds in between
-    folds.erase(folds.begin() + i);
-    foldSorted_ = false;
-    i--;
-  }
-}
-
-void TextEditor::updateFoldColumn(Coordinates& fold, const Coordinates& start,
-                                  const Coordinates& end) {
-  if (start.line >= text_.getTotalLines()) {
-    return;
-  }
-
-  const Line& startLine = text_.getLineGlyphs(start.line);
-  int colOffset = 0;
-  int charIndex = 0;
-  bool skipped = false;
-  const int bracketEndCharIndex = text_.getCharacterIndex(foldEnd_[&fold - &foldBegin_[0]]);
-
-  while (charIndex < static_cast<int>(startLine.size()) &&
-         (!skipped || (skipped && charIndex < bracketEndCharIndex))) {
-    char c = startLine[charIndex].character;
-    charIndex += Utf8::SequenceLength(c);
-
-    if (c == '\t') {
-      colOffset = (colOffset / tabSize_) * tabSize_ + tabSize_;
-    } else {
-      colOffset++;
-    }
-
-    // Move to the last line
-    if (charIndex == static_cast<int>(startLine.size()) && end.line < text_.getTotalLines() &&
-        !skipped) {
-      charIndex = text_.getCharacterIndex(end);
-      skipped = true;
-    }
-  }
-
-  fold.column = colOffset;
+  core_.removeFolds(folds, start, end);
 }
 
 namespace {
@@ -2511,184 +2078,15 @@ void TextEditor::render(std::string_view title, const ImVec2& size, bool showBor
 }
 
 void TextEditor::setText(std::string_view text) {
-  // Clear existing state
-  foldBegin_.clear();
-  foldEnd_.clear();
-  foldSorted_ = false;
-  undoBuffer_.clear();
-  undoIndex_ = 0;
-
-  text_.setText(text);
-
-// TODO: Find a way to add text folds
-#if 0
-
-    // Track fold markers
-    if (c == '{') {
-      foldBegin_.emplace_back(static_cast<int>(text_.getTotalLines() - 1),
-                              static_cast<int>(lines_.back().size()));
-    } else if (c == '}') {
-      foldEnd_.emplace_back(static_cast<int>(text_.getTotalLines() - 1),
-                            static_cast<int>(lines_.back().size()));
-    }
-#endif
-
-  // Update editor state
-  textChanged_ = true;
-  scrollToTop_ = true;
-
-  detectIndentationStyle();
-  colorize();
-}
-
-void TextEditor::handleNewLine(UndoState& state, const Coordinates& coord, bool smartIndent) {
-  Line& line = text_.getLineGlyphsMutable(coord.line);
-  Line& newLine = insertLine(coord.line, coord.column);
-
-  // Auto indentation
-  size_t whitespaceSize = 0;
-  std::string added = "\n";
-  if (languageDefinition_.autoIndentation && smartIndent) {
-    for (size_t i = 0;
-         i < line.size() && IsAscii(line[i].character) && std::isblank(line[i].character); ++i) {
-      ++whitespaceSize;
-      newLine.insert(newLine.begin(), line[i]);
-      added.push_back(line[i].character);
-      // foldOffset += 1 + (line[i].character == '\t') * 3;
-    }
-  }
-
-  // Update cursor and state
-  setCursorPosition(
-      Coordinates(coord.line + 1, text_.getCharacterColumn(coord.line + 1, whitespaceSize)));
-
-  state.record.added = added;
-}
-void TextEditor::handleRegularCharacter(UndoState& state, const Coordinates& coord,
-                                        ImWchar character) {
-  // Determine where to insert the new character(s)
-  int charIndex = text_.getCharacterIndex(coord);
-  std::string added;
-
-  Line& line = text_.getLineGlyphsMutable(coord.line);
-
-  // Special handling for tab character if typed alone
-  if (character == '\t') {
-    if (insertSpaces_) {
-      // Insert `tabSize_` spaces
-      line.insert(line.begin() + charIndex, tabSize_, Glyph(' ', ColorIndex::Default));
-      added.append(tabSize_, ' ');
-    } else {
-      // Insert a single '\t' character
-      line.insert(line.begin() + charIndex, Glyph('\t', ColorIndex::Default));
-      added.push_back('\t');
-    }
-
-    // Move cursor right by tab size if spaces, else by 1
-    state_.cursorPosition.column += insertSpaces_ ? tabSize_ : 1;
-
-    // Record undo state
-    state.record.added = added;
-    scrollToCursor_ = true;
-    return;
-  }
-
-  // For other characters, convert to UTF-8
-  char buf[5] = {};
-  ImTextCharToUtf8(buf, character);
-  std::string_view chars(buf);
-
-  // Insert the character(s) at the current cursor position
-  for (char ch : chars) {
-    if (ch == '\t' && insertSpaces_) {
-      // If we get a tab and we're using spaces, insert tabSize_ spaces
-      for (int i = 0; i < tabSize_; i++) {
-        line.insert(line.begin() + charIndex++, Glyph(' ', ColorIndex::Default));
-      }
-      added.append(tabSize_, ' ');
-    } else {
-      line.insert(line.begin() + charIndex++, Glyph(ch, ColorIndex::Default));
-      added.push_back(ch);
-    }
-  }
-
-  // Update cursor position
-  // For a tab, move cursor by tab size if using spaces, otherwise by 1
-  const int advance = (chars[0] == '\t' && insertSpaces_) ? tabSize_ : (int)chars.size();
-  state_.cursorPosition = Coordinates(coord.line, coord.column + advance);
-
-  // Record what was added for undo
-  state.record.added = added;
-
-  // Ensure the cursor is visible
-  scrollToCursor_ = true;
+  core_.setText(text);
 }
 
 void TextEditor::enterCharacter(ImWchar character, bool shift) {
-  UndoState state;
-  state.record.before = state_;
-
-  // Handle selection
-  if (hasSelection()) {
-    if (character == '\t') {
-      handleMultiLineTab(state, shift);
-      return;
-    } else {
-      state.record.removed = getSelectedText();
-      state.record.removedStart = state_.selectionStart;
-      state.record.removedEnd = state_.selectionEnd;
-      deleteSelection();
-    }
-  }
-
-  state.insertPos = getActualCursorCoordinates();
-  state.record.addedStart = state.insertPos;
-
-  // Handle different character types
-  if (character == '\n') {
-    handleNewLine(state, state.insertPos, smartIndent_);
-  } else {
-    handleRegularCharacter(state, state.insertPos, character);
-  }
-
-  // Update autocomplete
-  if (activeAutocomplete_ && character <= 127 && (std::isalpha(character) || character == '_')) {
-    requestAutocomplete_ = true;
-    readyForAutocomplete_ = false;
-  }
-
-  // Update change tracking
-  updateChangeTracking();
-
-  // Finalize changes
-  textChanged_ = true;
-  if (onContentUpdate) {
-    onContentUpdate(this);
-  }
-
-  state.record.addedEnd = getActualCursorCoordinates();
-  state.record.after = state_;
-  addUndo(state.record);
-
-  colorize(state.insertPos.line - 1, 3);
-
-  // Ensure cursor is visible after entering a character
-  scrollToCursor_ = true;
-
-  // Handle special features
-  handleFunctionTooltip(character, state.insertPos);
-  handleBraceCompletion(character);
+  core_.enterCharacter(static_cast<char32_t>(character), shift);
 }
 
 void TextEditor::updateChangeTracking() {
-  if (!scrollbarMarkers_) {
-    return;
-  }
-
-  const int currentLine = state_.cursorPosition.line;
-  if (std::find(changedLines_.begin(), changedLines_.end(), currentLine) == changedLines_.end()) {
-    changedLines_.push_back(currentLine);
-  }
+  core_.updateChangeTracking();
 }
 
 void TextEditor::handleFunctionTooltip(ImWchar character, const Coordinates& curPos) {
@@ -2725,106 +2123,16 @@ void TextEditor::handleFunctionTooltip(ImWchar character, const Coordinates& cur
   }
 }
 
-void TextEditor::handleBraceCompletion(ImWchar character) {
-  if (!completeBraces_) {
-    return;
-  }
-
-  ImWchar closingChar = 0;
-  switch (character) {
-    case '{':
-      enterCharacter('\n', false);
-      closingChar = '}';
-      break;
-    case '(': closingChar = ')'; break;
-    case '[': closingChar = ']'; break;
-    default: return;
-  }
-
-  if (closingChar != 0) {
-    enterCharacter(closingChar, false);
-
-    auto cursorPos = state_.cursorPosition;
-    cursorPos.column--;
-    setCursorPosition(cursorPos);
-  }
-}
-
-void TextEditor::handleMultiLineTab(UndoState& state, bool shift) {
-  auto start = state_.selectionStart;
-  auto end = state_.selectionEnd;
-  const auto originalEnd = end;
-
-  if (start > end) {
-    std::swap(start, end);
-  }
-
-  // Adjust selection to full lines
-  start.column = 0;
-  if (end.column == 0 && end.line > 0) {
-    end.line--;
-  }
-
-  if (end.line >= text_.getTotalLines()) {
-    end.line = text_.getTotalLines() == 0 ? 0 : text_.getTotalLines() - 1;
-  }
-  end.column = text_.getLineMaxColumn(end.line);
-
-  // Store undo state
-  state.record.removedStart = start;
-  state.record.removedEnd = end;
-  state.record.removed = getText(start, end);
-
-  bool modified = false;
-  for (int i = start.line; i <= end.line; i++) {
-    Line& line = text_.getLineGlyphsMutable(i);
-    if (shift) {
-      if (!line.empty()) {
-        if (line.front().character == '\t') {
-          line.erase(line.begin());
-          modified = true;
-        } else {
-          for (int j = 0; j < tabSize_ && !line.empty() && line.front().character == ' '; j++) {
-            line.erase(line.begin());
-            modified = true;
-          }
-        }
-      }
-    } else {
-      if (insertSpaces_) {
-        line.insert(line.begin(), tabSize_, Glyph(' ', ColorIndex::Background));
-      } else {
-        line.insert(line.begin(), Glyph('\t', ColorIndex::Background));
-      }
-      modified = true;
-    }
-  }
-
-  if (modified) {
-    state.record.addedStart = Coordinates(start.line, text_.getCharacterColumn(start.line, 0));
-
-    // Update selection
-    Coordinates rangeEnd;
-    if (originalEnd.column != 0) {
-      end = Coordinates(end.line, text_.getLineMaxColumn(end.line));
-      rangeEnd = end;
-      state.record.added = getText(start, end);
-    } else {
-      end = Coordinates(originalEnd.line, 0);
-      rangeEnd = Coordinates(end.line - 1, text_.getLineMaxColumn(end.line - 1));
-      state.record.added = getText(start, rangeEnd);
-    }
-
-    state.record.addedEnd = rangeEnd;
-    state.record.after = state_;
-
-    state_.selectionStart = start;
-    state_.selectionEnd = end;
-    ensureCursorVisible();
-  }
-}
 void TextEditor::setColorizerEnabled(bool enabled) {
-  colorizerEnabled_ = enabled;
+  core_.setColorizerEnabled(enabled);
+}
+
+void TextEditor::handleEndOfLineDelete(Coordinates pos, UndoRecord& undo) {
+  core_.handleEndOfLineDelete(pos, undo);
+}
+
+void TextEditor::handleMidLineDelete(Coordinates pos, UndoRecord& undo) {
+  core_.handleMidLineDelete(pos, undo);
 }
 
 Coordinates TextEditor::getCorrectCursorPosition() {
@@ -2845,512 +2153,75 @@ Coordinates TextEditor::getCorrectCursorPosition() {
 }
 
 void TextEditor::setCursorPosition(const Coordinates& position) {
-  if (state_.cursorPosition != position) {
-    state_.cursorPosition = position;
-    cursorPositionChanged_ = true;
-    scrollToCursor_ = true;
-  }
+  core_.setCursorPosition(position);
 }
 
 void TextEditor::setSelectionStart(const Coordinates& position) {
-  state_.selectionStart = sanitizeCoordinates(position);
-
-  if (state_.selectionStart > state_.selectionEnd) {
-    std::swap(state_.selectionStart, state_.selectionEnd);
-  }
+  core_.setSelectionStart(position);
 }
 
 void TextEditor::setSelectionEnd(const Coordinates& position) {
-  state_.selectionEnd = sanitizeCoordinates(position);
-
-  if (state_.selectionStart > state_.selectionEnd) {
-    std::swap(state_.selectionStart, state_.selectionEnd);
-  }
+  core_.setSelectionEnd(position);
 }
 
 void TextEditor::insertText(std::string_view text, bool indent) {
-  if (text.empty()) {
-    return;
-  }
-
-  auto pos = getActualCursorCoordinates();
-  auto start = std::min(pos, state_.selectionStart);
-  int totalLines = pos.line - start.line;
-
-  totalLines += insertTextAt(pos, text.data(), indent);
-
-  setSelection(pos, pos);
-  setCursorPosition(pos);
-  colorize(start.line - 1, totalLines + 2);
+  core_.insertText(text, indent);
 }
 
 void TextEditor::deleteSelection() {
-  UTILS_RELEASE_ASSERT(state_.selectionEnd >= state_.selectionStart);
-
-  if (state_.selectionEnd == state_.selectionStart) {
-    return;
-  }
-
-  deleteRange(state_.selectionStart, state_.selectionEnd);
-
-  setSelection(state_.selectionStart, state_.selectionStart);
-  setCursorPosition(state_.selectionStart);
-  colorize(state_.selectionStart.line, 1);
+  core_.deleteSelection();
 }
 
 void TextEditor::moveUp(int amount, bool select) {
-  const auto oldPos = state_.cursorPosition;
-  state_.cursorPosition.line = std::max(0, state_.cursorPosition.line - amount);
-
-  if (oldPos != state_.cursorPosition) {
-    if (select) {
-      if (oldPos == interactiveStart_) {
-        interactiveStart_ = state_.cursorPosition;
-      } else if (oldPos == interactiveEnd_) {
-        interactiveEnd_ = state_.cursorPosition;
-      } else {
-        interactiveStart_ = state_.cursorPosition;
-        interactiveEnd_ = oldPos;
-      }
-    } else {
-      interactiveStart_ = interactiveEnd_ = state_.cursorPosition;
-    }
-
-    setSelection(interactiveStart_, interactiveEnd_);
-    scrollToCursor_ = true;
-  }
+  core_.moveUp(amount, select);
 }
 
 void TextEditor::moveDown(int amount, bool select) {
-  UTILS_RELEASE_ASSERT(state_.cursorPosition.column >= 0);
-
-  const auto oldPos = state_.cursorPosition;
-  state_.cursorPosition.line =
-      std::max(0, std::min(text_.getTotalLines() - 1, state_.cursorPosition.line + amount));
-
-  if (state_.cursorPosition != oldPos) {
-    if (select) {
-      if (oldPos == interactiveEnd_) {
-        interactiveEnd_ = state_.cursorPosition;
-      } else if (oldPos == interactiveStart_) {
-        interactiveStart_ = state_.cursorPosition;
-      } else {
-        interactiveStart_ = oldPos;
-        interactiveEnd_ = state_.cursorPosition;
-      }
-    } else {
-      interactiveStart_ = interactiveEnd_ = state_.cursorPosition;
-    }
-
-    setSelection(interactiveStart_, interactiveEnd_);
-    scrollToCursor_ = true;
-    scrollToCursor_ = true;
-    scrollToCursor_ = true;
-  }
-}
-namespace {
-
-/**
- * Check if a character is a UTF-8 continuation byte.
- */
-bool isUtfSequence(char c) {
-  return (c & 0xC0) == 0x80;
+  core_.moveDown(amount, select);
 }
 
-}  // namespace
-
-void TextEditor::moveLeft(int amount, bool select, bool /*wordMode*/) {
-  auto oldPos = state_.cursorPosition;
-
-  while (amount-- > 0) {
-    if (state_.cursorPosition.column > 0) {
-      // Move left within line
-      state_.cursorPosition.column--;
-    } else {
-      // Move to end of previous line if exists
-      if (state_.cursorPosition.line > 0) {
-        state_.cursorPosition.line--;
-        state_.cursorPosition.column = text_.getLineMaxColumn(state_.cursorPosition.line);
-      }
-    }
-  }
-
-  if (select) {
-    if (oldPos == interactiveStart_) {
-      interactiveStart_ = state_.cursorPosition;
-    } else if (oldPos == interactiveEnd_) {
-      interactiveEnd_ = state_.cursorPosition;
-    } else {
-      interactiveStart_ = state_.cursorPosition;
-      interactiveEnd_ = oldPos;
-    }
-  } else {
-    interactiveStart_ = interactiveEnd_ = state_.cursorPosition;
-  }
-
-  setSelection(interactiveStart_, interactiveEnd_);
-  scrollToCursor_ = true;
+void TextEditor::moveLeft(int amount, bool select, bool wordMode) {
+  core_.moveLeft(amount, select, wordMode);
 }
 
-void TextEditor::moveRight(int amount, bool select, bool /*wordMode*/) {
-  auto oldPos = state_.cursorPosition;
-
-  while (amount-- > 0) {
-    int maxCol = text_.getLineMaxColumn(state_.cursorPosition.line);
-    if (state_.cursorPosition.column < maxCol) {
-      // Move right within line
-      state_.cursorPosition.column++;
-    } else {
-      // Move to start of next line if exists
-      if (state_.cursorPosition.line < text_.getTotalLines() - 1) {
-        state_.cursorPosition.line++;
-        state_.cursorPosition.column = 0;
-      }
-    }
-  }
-
-  if (select) {
-    if (oldPos == interactiveEnd_) {
-      interactiveEnd_ = state_.cursorPosition;
-    } else if (oldPos == interactiveStart_) {
-      interactiveStart_ = state_.cursorPosition;
-    } else {
-      interactiveStart_ = oldPos;
-      interactiveEnd_ = state_.cursorPosition;
-    }
-  } else {
-    interactiveStart_ = interactiveEnd_ = state_.cursorPosition;
-  }
-
-  setSelection(interactiveStart_, interactiveEnd_);
-  scrollToCursor_ = true;
+void TextEditor::moveRight(int amount, bool select, bool wordMode) {
+  core_.moveRight(amount, select, wordMode);
 }
 
 void TextEditor::moveTop(bool select) {
-  const auto oldPos = state_.cursorPosition;
-  setCursorPosition(Coordinates(0, 0));
-
-  if (state_.cursorPosition != oldPos) {
-    if (select) {
-      interactiveEnd_ = oldPos;
-      interactiveStart_ = state_.cursorPosition;
-    } else {
-      interactiveStart_ = interactiveEnd_ = state_.cursorPosition;
-    }
-    setSelection(interactiveStart_, interactiveEnd_);
-  }
+  core_.moveTop(select);
 }
 
 void TextEditor::moveBottom(bool select) {
-  const auto oldPos = getCursorPosition();
-  const auto newPos = Coordinates(text_.getTotalLines() - 1, 0);
-  setCursorPosition(newPos);
-
-  if (select) {
-    interactiveStart_ = oldPos;
-    interactiveEnd_ = newPos;
-  } else {
-    interactiveStart_ = interactiveEnd_ = newPos;
-  }
-  setSelection(interactiveStart_, interactiveEnd_);
+  core_.moveBottom(select);
 }
 
 void TextEditor::moveHome(bool select) {
-  const auto oldPos = state_.cursorPosition;
-  setCursorPosition(Coordinates(state_.cursorPosition.line, 0));
-
-  if (state_.cursorPosition != oldPos) {
-    if (select) {
-      if (oldPos == interactiveStart_) {
-        interactiveStart_ = state_.cursorPosition;
-      } else if (oldPos == interactiveEnd_) {
-        interactiveEnd_ = state_.cursorPosition;
-      } else {
-        interactiveStart_ = state_.cursorPosition;
-        interactiveEnd_ = oldPos;
-      }
-    } else {
-      interactiveStart_ = interactiveEnd_ = state_.cursorPosition;
-    }
-    setSelection(interactiveStart_, interactiveEnd_);
-  }
+  core_.moveHome(select);
 }
 
 void TextEditor::moveEnd(bool select) {
-  const auto oldPos = state_.cursorPosition;
-  setCursorPosition(Coordinates(state_.cursorPosition.line, text_.getLineMaxColumn(oldPos.line)));
-
-  if (state_.cursorPosition != oldPos) {
-    if (select) {
-      if (oldPos == interactiveEnd_) {
-        interactiveEnd_ = state_.cursorPosition;
-      } else if (oldPos == interactiveStart_) {
-        interactiveStart_ = state_.cursorPosition;
-      } else {
-        interactiveStart_ = oldPos;
-        interactiveEnd_ = state_.cursorPosition;
-      }
-    } else {
-      interactiveStart_ = interactiveEnd_ = state_.cursorPosition;
-    }
-    setSelection(interactiveStart_, interactiveEnd_);
-  }
-}
-
-void TextEditor::handleEndOfLineDelete(Coordinates pos, UndoRecord& undo) {
-  if (pos.line == text_.getTotalLines() - 1) {
-    return;
-  }
-
-  Line& line = text_.getLineGlyphsMutable(pos.line);
-  undo.removed = "\n";
-  undo.removedStart = undo.removedEnd = getActualCursorCoordinates();
-  advance(undo.removedEnd);
-
-  // Move folds
-  // for (auto &fold : getFoldBegin()) {
-  //   if (fold.line == pos.line + 1) {
-  //     fold.line = std::max(0, fold.line - 1);
-  //     fold.column += getCharacterColumn(pos.line, line.size());
-  //   }
-  // }
-
-  // for (auto &fold : getFoldEnd()) {
-  //   if (fold.line == pos.line + 1) {
-  //     fold.line = std::max(0, fold.line - 1);
-  //     fold.column += getCharacterColumn(pos.line, line.size());
-  //   }
-  // }
-
-  // Merge lines
-  const Line& nextLine = text_.getLineGlyphs(pos.line + 1);
-  line.insert(line.end(), nextLine.begin(), nextLine.end());
-  removeLine(pos.line + 1);
-}
-
-void TextEditor::handleMidLineDelete(Coordinates pos, UndoRecord& undo) {
-  Line& line = text_.getLineGlyphsMutable(pos.line);
-  auto charIndex = text_.getCharacterIndex(pos);
-
-  undo.removedStart = undo.removedEnd = getActualCursorCoordinates();
-  undo.removedEnd.column++;
-  undo.removed = getText(undo.removedStart, undo.removedEnd);
-
-  removeFolds(undo.removedStart, undo.removedEnd);
-
-  auto charLen = Utf8::SequenceLength(line[charIndex].character);
-  while (charLen-- > 0 && charIndex < static_cast<int>(line.size())) {
-    line.erase(line.begin() + charIndex);
-  }
+  core_.moveEnd(select);
 }
 
 void TextEditor::delete_() {
-  if (text_.getTotalLines() == 0) {
-    return;
-  }
-
-  UndoRecord undo;
-  undo.before = state_;
-
-  if (hasSelection()) {
-    undo.removed = getSelectedText();
-    undo.removedStart = state_.selectionStart;
-    undo.removedEnd = state_.selectionEnd;
-    deleteSelection();
-  } else {
-    auto pos = getActualCursorCoordinates();
-    setCursorPosition(pos);
-
-    if (pos.column == text_.getLineMaxColumn(pos.line)) {
-      handleEndOfLineDelete(pos, undo);
-    } else {
-      handleMidLineDelete(pos, undo);
-    }
-
-    updateChangeTracking();
-    textChanged_ = true;
-
-    if (onContentUpdate) {
-      onContentUpdate(this);
-    }
-
-    colorize(pos.line, 1);
-  }
-
-  undo.after = state_;
-  addUndo(undo);
-}
-
-void TextEditor::handleStartOfLineDelete(const Coordinates& pos, UndoRecord& undo) {
-  if (pos.line == 0) {
-    return;
-  }
-
-  undo.removed = "\n";
-  undo.removedStart = undo.removedEnd =
-      Coordinates(pos.line - 1, text_.getLineMaxColumn(pos.line - 1));
-  advance(undo.removedEnd);
-
-  const Line& line = text_.getLineGlyphs(pos.line);
-  Line& prevLine = text_.getLineGlyphsMutable(pos.line - 1);
-  const int prevSize = text_.getLineMaxColumn(pos.line - 1);
-
-  // Merge lines
-  prevLine.insert(prevLine.end(), line.begin(), line.end());
-
-  // Update error markers
-  ErrorMarkers updatedMarkers;
-  for (const auto& marker : errorMarkers_) {
-    updatedMarkers.insert(
-        {marker.first - 1 == pos.line ? marker.first - 1 : marker.first, marker.second});
-  }
-  setErrorMarkers(updatedMarkers);
-
-  // // Update fold positions
-  // for (auto &fold : getFoldBegin()) {
-  //   if (fold.line == pos.line) {
-  //     fold.line = std::max(0, fold.line - 1);
-  //     fold.column += prevSize;
-  //   }
-  // }
-
-  // for (auto &fold : getFoldEnd()) {
-  //   if (fold.line == pos.line) {
-  //     fold.line = std::max(0, fold.line - 1);
-  //     fold.column += prevSize;
-  //   }
-  // }
-
-  removeLine(pos.line);
-  setCursorPosition(Coordinates(pos.line - 1, prevSize));
-}
-
-void TextEditor::handleMidLineBackspace(const Coordinates& pos, UndoRecord& undo) {
-  // Before removing chars, detect if we're at an indentation boundary.
-  // If insertSpaces_ is true and we're in leading spaces, remove tabSize_
-  // spaces at once if possible. If insertSpaces_ is false and the previous char
-  // is '\t', remove just the tab.
-
-  if (pos.column > 0 && pos.column <= tabSize_) {
-    // Possibly at the start of a block indentation
-    Line& line = text_.getLineGlyphsMutable(pos.line);
-    int charIndex = text_.getCharacterIndex(pos);
-    // Look back one indentation unit
-    if (insertSpaces_) {
-      // Check if the previous tabSize_ chars are spaces
-      int startIndex = charIndex - 1;
-      int countSpaces = 0;
-      for (int i = 0; i < tabSize_ && startIndex - i >= 0; i++) {
-        if (line[startIndex - i].character == ' ')
-          countSpaces++;
-        else
-          break;
-      }
-      if (countSpaces == tabSize_) {
-        // Remove a whole indentation unit of spaces
-        undo.removedStart = Coordinates(pos.line, pos.column - tabSize_);
-        undo.removedEnd = Coordinates(pos.line, pos.column);
-        undo.removed = getText(undo.removedStart, undo.removedEnd);
-        line.erase(line.begin() + (charIndex - tabSize_), line.begin() + charIndex);
-        setCursorPosition(Coordinates(pos.line, pos.column - tabSize_));
-        return;
-      }
-    } else {
-      // If using tabs, check if previous character is a '\t'
-      int prevIndex = charIndex - 1;
-      if (prevIndex >= 0 && line[prevIndex].character == '\t') {
-        undo.removedStart = Coordinates(pos.line, pos.column - 1);
-        undo.removedEnd = Coordinates(pos.line, pos.column);
-        undo.removed = "\t";
-        line.erase(line.begin() + prevIndex);
-        setCursorPosition(Coordinates(pos.line, pos.column - 1));
-        return;
-      }
-    }
-  }
-
-  Line& line = text_.getLineGlyphsMutable(pos.line);
-
-  // Find the character cluster to remove
-  int charIndex = text_.getCharacterIndex(pos) - 1;
-  while (charIndex > 0 && isUtfSequence(line[charIndex].character)) {
-    --charIndex;
-  }
-
-  int charLen = Utf8::SequenceLength(line[charIndex].character);
-  int endIndex = charIndex + charLen;
-
-  // Calculate column positions before and after the character
-  int preColumn = text_.getCharacterColumn(pos.line, charIndex);
-  int postColumn = text_.getCharacterColumn(pos.line, endIndex);
-  int remSize = postColumn - preColumn;
-
-  // Set undo information
-  undo.removedStart = Coordinates(pos.line, preColumn);
-  undo.removedEnd = Coordinates(pos.line, postColumn);
-  undo.removed = getText(undo.removedStart, undo.removedEnd);
-
-  // Remove the character cluster
-  line.erase(line.begin() + charIndex, line.begin() + endIndex);
-
-  // Move cursor back by exactly remSize columns
-  setCursorPosition(Coordinates(pos.line, pos.column - remSize));
+  core_.delete_();
 }
 
 void TextEditor::backspace() {
-  if (text_.getTotalLines() == 0) {
-    return;
-  }
-
-  UndoRecord undo;
-  undo.before = state_;
-
-  if (hasSelection()) {
-    undo.removed = getSelectedText();
-    undo.removedStart = state_.selectionStart;
-    undo.removedEnd = state_.selectionEnd;
-    deleteSelection();
-  } else {
-    auto pos = getActualCursorCoordinates();
-    setCursorPosition(pos);
-
-    if (state_.cursorPosition.column == 0) {
-      handleStartOfLineDelete(pos, undo);
-    } else {
-      handleMidLineBackspace(pos, undo);
-    }
-
-    updateChangeTracking();
-    textChanged_ = true;
-
-    if (onContentUpdate) {
-      onContentUpdate(this);
-    }
-
-    ensureCursorVisible();
-    colorize(state_.cursorPosition.line, 1);
-  }
-
-  undo.after = state_;
-  addUndo(undo);
-
-  if (activeAutocomplete_ && autocompleteOpened_) {
-    requestAutocomplete_ = true;
-    readyForAutocomplete_ = false;
-  }
+  core_.backspace();
 }
 
 void TextEditor::selectWordUnderCursor() {
-  const auto cursorPos = getCursorPosition();
-  setSelection(findWordStart(cursorPos), findWordEnd(cursorPos));
+  core_.selectWordUnderCursor();
 }
 
 void TextEditor::selectAll() {
-  setSelection(Coordinates(0, 0), Coordinates(text_.getTotalLines(), 0));
+  core_.selectAll();
 }
 
 bool TextEditor::hasSelection() const {
-  return state_.selectionEnd > state_.selectionStart;
+  return core_.hasSelection();
 }
 
 void TextEditor::setShortcut(ShortcutId id, Shortcut shortcut) {
@@ -3420,23 +2291,19 @@ void TextEditor::paste() {
 }
 
 bool TextEditor::canUndo() const {
-  return undoIndex_ > 0;
+  return core_.canUndo();
 }
 
 bool TextEditor::canRedo() const {
-  return undoIndex_ < static_cast<int>(undoBuffer_.size());
+  return core_.canRedo();
 }
 
 void TextEditor::undo(int steps) {
-  while (canUndo() && steps-- > 0) {
-    undoBuffer_[--undoIndex_].undo(this);
-  }
+  core_.undo(steps);
 }
 
 void TextEditor::redo(int steps) {
-  while (canRedo() && steps-- > 0) {
-    undoBuffer_[undoIndex_++].redo(this);
-  }
+  core_.redo(steps);
 }
 
 const TextEditor::Palette& TextEditor::getDarkPalette() {
@@ -3477,232 +2344,23 @@ const TextEditor::Palette& TextEditor::getDarkPalette() {
 }
 
 std::string TextEditor::getText() const {
-  return text_.getText();
+  return core_.getText();
 }
 
 std::string TextEditor::getSelectedText() const {
-  return getText(state_.selectionStart, state_.selectionEnd);
+  return core_.getSelectedText();
 }
 
 void TextEditor::colorize(int fromLine, int lines) {
-  const int toLine =
-      lines == -1 ? text_.getTotalLines() : std::min<int>(text_.getTotalLines(), fromLine + lines);
-
-  colorRangeMin_ = std::min(colorRangeMin_, fromLine);
-  colorRangeMax_ = std::max(colorRangeMax_, toLine);
-  colorRangeMin_ = std::max(0, colorRangeMin_);
-  colorRangeMax_ = std::max(colorRangeMin_, colorRangeMax_);
-  checkComments_ = true;
+  core_.colorize(fromLine, lines);
 }
 
 void TextEditor::colorizeRange(int fromLine, int toLine) {
-  if (text_.getTotalLines() == 0 || !colorizerEnabled_) {
-    return;
-  }
-
-  std::string buffer;
-  std::cmatch results;
-  std::string_view identifier;
-
-  const int endLine = std::max(0, std::min(text_.getTotalLines(), toLine));
-
-  for (int i = fromLine; i < endLine; ++i) {
-    Line& line = text_.getLineGlyphsMutable(i);
-    if (line.empty()) {
-      continue;
-    }
-
-    // Build buffer from line
-    buffer.resize(line.size());
-    for (size_t j = 0; j < line.size(); ++j) {
-      buffer[j] = line[j].character;
-      line[j].colorIndex = ColorIndex::Default;
-    }
-
-    const char* current = buffer.data();
-    const char* end = current + buffer.size();
-
-    while (current != end) {
-      const char* tokenBegin = nullptr;
-      const char* tokenEnd = nullptr;
-      ColorIndex tokenColor = ColorIndex::Default;
-      bool hasToken = false;
-
-      // Try custom tokenizer if available
-      if (languageDefinition_.tokenize) {
-        if (languageDefinition_.tokenize(current, end, tokenBegin, tokenEnd, tokenColor)) {
-          hasToken = true;
-        }
-      }
-
-      // Try regex matching if no custom token found
-      if (!hasToken) {
-        for (const auto& [regex, color] : regexList_) {
-          if (std::regex_search(current, end, results, regex,
-                                std::regex_constants::match_continuous)) {
-            const auto& match = *results.begin();
-            tokenBegin = match.first;
-            tokenEnd = match.second;
-            tokenColor = color;
-            hasToken = true;
-            break;
-          }
-        }
-      }
-
-      if (!hasToken) {
-        ++current;
-        continue;
-      }
-
-      // Process token
-      const size_t tokenLength = tokenEnd - tokenBegin;
-      if (tokenColor == ColorIndex::Identifier) {
-        identifier = std::string_view(tokenBegin, tokenLength);
-
-        // Handle case sensitivity
-        std::string upperIdentifier;
-        if (!languageDefinition_.caseSensitive) {
-          upperIdentifier.resize(identifier.size());
-          std::transform(identifier.begin(), identifier.end(), upperIdentifier.begin(), ::toupper);
-          identifier = upperIdentifier;
-        }
-
-        if (languageDefinition_.keywords.count(std::string(identifier))) {
-          tokenColor = ColorIndex::Keyword;
-        } else if (languageDefinition_.identifiers.count(std::string(identifier))) {
-          tokenColor = ColorIndex::KnownIdentifier;
-        }
-      }
-
-      // Apply color to token
-      const size_t offset = tokenBegin - buffer.data();
-      for (size_t j = 0; j < tokenLength; ++j) {
-        line[offset + j].colorIndex = tokenColor;
-      }
-
-      current = tokenEnd;
-    }
-  }
+  core_.colorizeRange(fromLine, toLine);
 }
-
-namespace {
-
-bool matchesCommentStart(std::string_view line, size_t index, std::string_view commentStart) {
-  if (index + commentStart.size() > line.size()) {
-    return false;
-  }
-  return std::equal(commentStart.begin(), commentStart.end(), line.begin() + index);
-}
-
-bool matchesCommentEnd(std::string_view line, size_t index, std::string_view commentEnd) {
-  if (index + 1 < commentEnd.size()) {
-    return false;
-  }
-  return std::equal(commentEnd.begin(), commentEnd.end(),
-                    line.begin() + index + 1 - commentEnd.size());
-}
-
-}  // namespace
 
 void TextEditor::colorizeInternal() {
-  if (text_.getTotalLines() == 0 || !colorizerEnabled_) {
-    return;
-  }
-
-  if (checkComments_) {
-    const auto endLine = text_.getTotalLines();
-    size_t commentStartLine = endLine;
-    size_t commentStartIndex = 0;
-    bool withinString = false;
-    bool withinSingleLineComment = false;
-    bool concatenate = false;  // '\' at end of line
-
-    for (size_t currentLine = 0; currentLine < endLine;) {
-      Line& line = text_.getLineGlyphsMutable(currentLine);
-      if (line.empty()) {
-        currentLine++;
-        continue;
-      }
-
-      for (size_t currentIndex = 0; currentIndex < line.size();) {
-        if (currentIndex == 0 && !concatenate) {
-          withinSingleLineComment = false;
-        }
-
-        auto& glyph = line[currentIndex];
-        const char c = glyph.character;
-
-        concatenate = (currentIndex == line.size() - 1 && c == '\\');
-
-        const bool inComment =
-            (commentStartLine < currentLine ||
-             (commentStartLine == currentLine && commentStartIndex <= currentIndex));
-
-        if (withinString) {
-          glyph.isMultiLineComment = inComment;
-
-          if (c == '\"') {
-            if (currentIndex + 1 < line.size() && line[currentIndex + 1].character == '\"') {
-              currentIndex++;
-              if (currentIndex < line.size()) {
-                line[currentIndex].isMultiLineComment = inComment;
-              }
-            } else {
-              withinString = false;
-            }
-          } else if (c == '\\') {
-            currentIndex++;
-            if (currentIndex < line.size()) {
-              line[currentIndex].isMultiLineComment = inComment;
-            }
-          }
-        } else {
-          if (c == '\"') {
-            withinString = true;
-            glyph.isMultiLineComment = inComment;
-          } else {
-            const std::string_view lineView(reinterpret_cast<const char*>(&line[0].character),
-                                            line.size());
-
-            if (!withinSingleLineComment &&
-                matchesCommentStart(lineView, currentIndex, languageDefinition_.commentStart)) {
-              commentStartLine = currentLine;
-              commentStartIndex = currentIndex;
-            } else if (!languageDefinition_.singleLineComment.empty() &&
-                       matchesCommentStart(lineView, currentIndex,
-                                           languageDefinition_.singleLineComment)) {
-              withinSingleLineComment = true;
-            }
-
-            glyph.isMultiLineComment = inComment;
-            glyph.isComment = withinSingleLineComment;
-
-            if (matchesCommentEnd(lineView, currentIndex, languageDefinition_.commentEnd)) {
-              commentStartLine = endLine;
-              commentStartIndex = 0;
-            }
-          }
-        }
-
-        currentIndex += Utf8::SequenceLength(c);
-      }
-      currentLine++;
-    }
-    checkComments_ = false;
-  }
-
-  if (colorRangeMin_ < colorRangeMax_) {
-    const int increment = (languageDefinition_.tokenize == nullptr) ? 10 : 10000;
-    const int to = std::min(colorRangeMin_ + increment, colorRangeMax_);
-    colorizeRange(colorRangeMin_, to);
-    colorRangeMin_ = to;
-
-    if (colorRangeMax_ == colorRangeMin_) {
-      colorRangeMin_ = std::numeric_limits<int>::max();
-      colorRangeMax_ = 0;
-    }
-  }
+  core_.colorizeInternal();
 }
 
 float TextEditor::getTextDistanceToLineStart(const Coordinates& pos) const {
@@ -3777,52 +2435,12 @@ int TextEditor::getPageSize() const {
   return static_cast<int>(std::floor(height / charAdvance_.y));
 }
 
-UndoRecord::UndoRecord(std::string_view added, const Coordinates& addedStart,
-                       const Coordinates& addedEnd, std::string_view removed,
-                       const Coordinates& removedStart, const Coordinates& removedEnd,
-                       const EditorState& before, const EditorState& after)
-    : added(added),
-      addedStart(addedStart),
-      addedEnd(addedEnd),
-      removed(removed),
-      removedStart(removedStart),
-      removedEnd(removedEnd),
-      before(before),
-      after(after) {
-  UTILS_RELEASE_ASSERT(addedStart <= addedEnd);
-  UTILS_RELEASE_ASSERT(removedStart <= removedEnd);
-}
+// `UndoRecord::UndoRecord`, `UndoRecord::undo`, and `UndoRecord::redo`
+// moved to TextEditorCore.cc as part of the C1 extraction. They now
+// operate on a `TextEditorCore*` rather than a `TextEditor*`.
 
-void UndoRecord::undo(TextEditor* editor) {
-  if (!added.empty()) {
-    editor->deleteRange(addedStart, addedEnd);
-    editor->colorize(addedStart.line - 1, addedEnd.line - addedStart.line + 2);
-  }
-
-  if (!removed.empty()) {
-    auto start = removedStart;
-    editor->insertTextAt(start, removed);
-    editor->colorize(removedStart.line - 1, removedEnd.line - removedStart.line + 2);
-  }
-
-  editor->state_ = before;
-}
-
-void UndoRecord::redo(TextEditor* editor) {
-  if (!removed.empty()) {
-    editor->deleteRange(removedStart, removedEnd);
-    editor->colorize(removedStart.line - 1, removedEnd.line - removedStart.line + 1);
-  }
-
-  if (!added.empty()) {
-    auto start = addedStart;
-    editor->insertTextAt(start, added);
-    editor->colorize(addedStart.line - 1, addedEnd.line - addedStart.line + 1);
-  }
-
-  editor->state_ = after;
-}
-
+// `LanguageDefinition::SVG()` moved to TextEditorCore.cc.
+#if 0
 const TextEditor::LanguageDefinition& TextEditor::LanguageDefinition::SVG() {
   static const LanguageDefinition langDef = [] {
     LanguageDefinition def;
@@ -4061,70 +2679,10 @@ const TextEditor::LanguageDefinition& TextEditor::LanguageDefinition::SVG() {
 
   return langDef;
 }
+#endif
 
 void TextEditor::detectIndentationStyle() {
-  int tabCount = 0;
-  int spaceCount = 0;
-  std::vector<int> indentLevels;
-  int linesChecked = std::min<int>(text_.getTotalLines(), 50);
-
-  for (int i = 0; i < linesChecked; ++i) {
-    const Line& line = text_.getLineGlyphs(i);
-    int leadingSpaces = 0;
-    int leadingTabs = 0;
-    for (auto& g : line) {
-      if (g.character == ' ')
-        leadingSpaces++;
-      else if (g.character == '\t')
-        leadingTabs++;
-      else
-        break;
-    }
-
-    // Count total spaces/tabs to guess mode
-    spaceCount += leadingSpaces;
-    tabCount += leadingTabs;
-
-    // Collect indentation levels for guessing tab size if using spaces
-    if (leadingTabs == 0 && leadingSpaces > 0) {
-      indentLevels.push_back(leadingSpaces);
-    }
-  }
-
-  // Decide indent mode
-  if (tabCount > spaceCount) {
-    indentMode_ = IndentMode::Tabs;
-    insertSpaces_ = false;
-  } else {
-    indentMode_ = IndentMode::Spaces;
-    insertSpaces_ = true;
-
-    // If we have indentation levels, find gcd for tab size guess
-    if (!indentLevels.empty()) {
-      auto gcd = [](int a, int b) {
-        while (b != 0) {
-          int t = a % b;
-          a = b;
-          b = t;
-        }
-        return a;
-      };
-
-      int computedTabSize = indentLevels[0];
-      for (size_t i = 1; i < indentLevels.size(); ++i) {
-        computedTabSize = gcd(computedTabSize, indentLevels[i]);
-        if (computedTabSize < 2) {
-          // Too small to be meaningful, default to 4
-          computedTabSize = 4;
-          break;
-        }
-      }
-      tabSize_ = std::max(1, computedTabSize);
-    } else {
-      // Fallback if no indent info
-      tabSize_ = 4;
-    }
-  }
+  core_.detectIndentationStyle();
 }
 
 }  // namespace donner::editor
